Move year-month-inputs logic into utility helper and incorporate it where needed.
