+++ src/add/debt.njk
@@ -53,6 +53,5 @@
 {% endblock %}
 
 {% block scripts %}
-  <script src="/scripts/helpers/year-month-inputs.js"></script>
   <script src="/scripts/add/debt.mjs" type="module"></script>
 {% endblock %}
