+++ src/add/savings.njk
@@ -42,6 +42,5 @@
 {% endblock %}
 
 {% block scripts %}
-  <script src="/scripts/helpers/year-month-inputs.js"></script>
   <script src="/scripts/add/savings.mjs" type="module"></script>
 {% endblock %}
