+++ src/edit/debt.njk
@@ -73,6 +73,5 @@
 {% endblock %}
 
 {% block scripts %}
-  <script src="/scripts/helpers/year-month-inputs.js"></script>
   <script src="/scripts/edit/debt.mjs" type="module"></script>
 {% endblock %}
