+++ src/edit/savings.njk
@@ -62,6 +62,5 @@
 {% endblock %}
 
 {% block scripts %}
-  <script src="/scripts/helpers/year-month-inputs.js"></script>
   <script src="/scripts/edit/savings.mjs" type="module"></script>
 {% endblock %}
